requirements.txt
nltk
mlflow
apache-airflow
flask
pandas
scikit-learn
ipykernel
python-box
pyyaml
psycopg2
prometheus_client
psycopg2-binary
-e .